[14.0][UPD] rename  pos_fixed_discount_in_lines -> pos_multi_discount_global
