--- pos_fixed_discount_in_lines/__init__.py
@@ -1,1 +1,0 @@
-from . import models
