--- pos_fixed_discount_in_lines/data/product.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <data noupdate="0">
-
-        <record id="product_fixed_discount_rounding" model="product.product">
-            <field name="name">Fixed discount rounding</field>
-            <field name="available_in_pos">True</field>
-            <field name="standard_price">0.00</field>
-            <field name="list_price">0.00</field>
-            <field name="weight">0.00</field>
-            <field name="type">consu</field>
-            <field name="categ_id" ref="point_of_sale.product_category_pos" />
-            <field name="uom_id" ref="uom.product_uom_unit" />
-            <field name="uom_po_id" ref="uom.product_uom_unit" />
-            <field name="default_code">FDR</field>
-            <field name="purchase_ok">False</field>
-            <field name="taxes_id" eval="False" />
-        </record>
-
-    </data>
-</odoo>
