--- pos_fixed_discount_in_lines/models/__init__.py
@@ -1,2 +1,0 @@
-from . import pos_config
-from . import pos_order
