--- pos_fixed_discount_in_lines/readme/CONFIGURATION.rst
@@ -1,1 +1,0 @@
-Go to POS config > activate "Split Fixed Discount Among Lines" and select a rounding product.
