--- pos_fixed_discount_in_lines/readme/CONTRIBUTORS.rst
@@ -1,3 +1,0 @@
-* Ooops404 <https://ooops404.com>
-
-  * Ilyas <irazor147@gmail.com>
