--- pos_fixed_discount_in_lines/readme/USAGE.rst
@@ -1,7 +1,0 @@
-- Fixed discount is applied on payment screen of POS interface, in order to make sure the discount is applied to all products in the order.
-
-.. figure:: ../static/description/payment.png
-
-- Once validated, in POS > Orders you can check the fixed discount applied to the order, the discount % on each line, and the necessary rounding amount.
-
-.. figure:: ../static/description/order.png
