--- pos_fixed_discount_in_lines/static/src/css/pos.css
@@ -1,18 +1,0 @@
-.payment-screen .discount-controls {
-    margin-top: 10px;
-}
-
-.payment-screen .paymentlines-empty .fixed_discount_line {
-    text-align: center;
-    font-size: 20px;
-    margin-bottom: 10px;
-}
-
-.payment-screen .fixed_discount_label {
-    font-size: 16px;
-    padding-right: 7px;
-}
-
-.payment-screen .fixed_discount_amount {
-    font-size: 16px;
-}
