--- pos_fixed_discount_in_lines/static/src/js/OrderWidget.js
@@ -1,27 +1,0 @@
-odoo.define("pos_fixed_discount_in_lines.PosOrderWidgetMod", function (require) {
-    "use strict";
-
-    const Registries = require("point_of_sale.Registries");
-    const {useState} = owl.hooks;
-    var OrderWidget = require("point_of_sale.OrderWidget");
-
-    const PosOrderWidgetMod = (OrderWidget) =>
-        class extends OrderWidget {
-            constructor() {
-                super(...arguments);
-                this.state = useState({
-                    total: 0,
-                    tax: 0,
-                    fixed_discount: 0,
-                    subtotal: 0,
-                });
-                this._updateSummary();
-            }
-
-            _updateSummary() {
-                super._updateSummary(...arguments);
-            }
-        };
-    Registries.Component.extend(OrderWidget, PosOrderWidgetMod);
-    return OrderWidget;
-});
