--- pos_fixed_discount_in_lines/static/src/xml/Screens/PaymentScreen/PaymentScreen.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<templates id="template" xml:space="preserve">
-
-    <t t-inherit="point_of_sale.PaymentScreen" t-inherit-mode="extension" owl="1">
-        <xpath expr="//div[hasclass('payment-controls')]" position="after">
-            <div class="discount-controls">
-                <div
-                    t-if="env.pos.config.split_fixed_discount"
-                    class="button js_fixed_discount"
-                    t-att-class="{ highlight: currentOrder.get_fixed_discount_enabled() }"
-                    t-on-click="toggleDiscountEnabled"
-                >
-                    <i class="fa fa-tag" />
-                    Discount (fixed)
-                </div>
-            </div>
-        </xpath>
-        <xpath
-            expr="//div[hasclass('right-content')]//PaymentScreenStatus"
-            position="replace"
-        >
-            <PaymentScreenStatus
-                paymentLines="paymentLines"
-                fixed_discount="state.fixed_discount"
-            />
-        </xpath>
-    </t>
-
-</templates>
