--- pos_fixed_discount_in_lines/static/src/xml/Screens/PaymentScreen/PaymentScreenStatus.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<templates id="template" xml:space="preserve">
-
-    <t t-inherit="point_of_sale.PaymentScreenStatus" t-inherit-mode="extension" owl="1">
-        <xpath expr="//div[hasclass('payment-status-remaining')]" position="after">
-            <div t-if="props.fixed_discount">
-                <span class="fixed_discount_label">Fixed Discount</span>
-                <span class="fixed_discount_amount">
-                    <t t-esc="props.fixed_discount" />
-                </span>
-            </div>
-        </xpath>
-        <xpath expr="//div[hasclass('total')]" position="after">
-            <div t-if="props.fixed_discount" class="fixed_discount_line">
-                Fixed discount applied: <t t-esc="props.fixed_discount" />
-            </div>
-        </xpath>
-    </t>
-</templates>
