--- pos_fixed_discount_in_lines/static/src/xml/Screens/ProductScreen/OrderWidget.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<templates id="template" xml:space="preserve">
-
-    <t t-inherit="point_of_sale.OrderWidget" t-inherit-mode="extension" owl="1">
-        <xpath expr="//OrderSummary" position="replace">
-            <OrderSummary
-                total="state.total"
-                tax="state.tax"
-                fixed_discount="state.fixed_discount"
-                subtotal="state.subtotal"
-            />
-        </xpath>
-    </t>
-
-</templates>
