--- pos_fixed_discount_in_lines/static/src/xml/Screens/ProductScreen/Orderline.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<templates id="template" xml:space="preserve">
-    <t t-inherit="point_of_sale.Orderline" t-inherit-mode="extension" owl="1">
-        <xpath expr="//li[@id='pos_line_manual_discount']/.." position="after">
-            <t t-if="props.line.get_fixed_discount() !== 0">
-                <li class="info">
-                    Fixed discount
-                    <em>
-                        <t t-esc="props.line.get_fixed_discount()" />%
-                    </em>
-                </li>
-            </t>
-        </xpath>
-    </t>
-</templates>
