--- pos_fixed_discount_in_lines/views/pos_discount_views.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="pos_config_view_form_inherit_pos_discount_inherit" model="ir.ui.view">
-        <field name="name">pos.config.form.inherit.pos.discount.inherit</field>
-        <field name="model">pos.config</field>
-        <field
-            name="inherit_id"
-            ref="pos_discount.pos_config_view_form_inherit_pos_discount"
-        />
-        <field name="arch" type="xml">
-            <xpath expr="////field[@name='module_pos_discount']/../.." position="after">
-                <div class="col-xs-12 col-lg-6 o_setting_box">
-                    <div class="o_setting_left_pane" />
-                    <div class="o_setting_right_pane">
-                        <div class="content-group">
-                            <field name="split_fixed_discount" />
-                            <label for="split_fixed_discount" />
-                        </div>
-                        <div
-                            class="content-group"
-                            attrs="{'invisible':[('split_fixed_discount','=',False)]}"
-                        >
-                            <field name="only_positive_discount" />
-                            <label for="only_positive_discount" />
-                        </div>
-                        <div
-                            class="row"
-                            attrs="{'invisible':[('split_fixed_discount','=',False)]}"
-                        >
-                            <label
-                                string="Rounding Product"
-                                for="rounding_product_id"
-                                class="col-lg-4 o_light_label"
-                            />
-                            <field
-                                name="rounding_product_id"
-                                attrs="{'required':[('split_fixed_discount','=',True)]}"
-                            />
-                        </div>
-                    </div>
-                </div>
-            </xpath>
-        </field>
-    </record>
-</odoo>
