--- pos_fixed_discount_in_lines/views/pos_views.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="pos_order_form_inherit" model="ir.ui.view">
-        <field name="name">pos.order.form.inherit</field>
-        <field name="model">pos.order</field>
-        <field name="inherit_id" ref="point_of_sale.view_pos_pos_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='amount_tax']" position='after'>
-                <field name="total_fixed_discount" widget="monetary" />
-            </xpath>
-            <xpath
-                expr="//field[@name='lines']//tree//field[@name='discount']"
-                position='before'
-            >
-                <field name="fixed_discount" invisible="1" />
-                <field name="fixed_discount_relative_amount" optional="show" />
-                <field
-                    name="fixed_discount_relative"
-                    string="Fixed Disc.%"
-                    optional="hide"
-                />
-            </xpath>
-            <xpath
-                expr="//field[@name='lines']//tree//field[@name='discount']"
-                position='attributes'
-            >
-                <attribute name="string">Total Discount %</attribute>
-            </xpath>
-        </field>
-    </record>
-</odoo>
