+++ pos_multi_discount_global/static/src/js/OrderWidget.js
@@ -13,8 +13,6 @@ odoo.define("pos_multi_discount_global.PosOrderWidgetMod", function (require) {
                     total: 0,
                     tax: 0,
                     fixed_discount: 0,
-                    percent_discount: 0,
-                    percent_discount_amount: 0,
                     subtotal: 0,
                 });
                 this._updateSummary();
