+++ pos_multi_discount_global/static/src/js/PaymentScreen.js
@@ -9,19 +9,15 @@ odoo.define("pos_multi_discount_global.PaymentScreen", function (require) {
     var utils = require("web.utils");
     var round_pr = utils.round_precision;
 
-    const MultiDiscountPaymentScreen = (PaymentScreen) =>
+    const FixedDiscountPaymentScreen = (PaymentScreen) =>
         class extends PaymentScreen {
             constructor() {
                 super(...arguments);
                 this.state = useState({
                     fixed_discount: 0,
-                    percent_discount: 0,
-                    percent_discount_amount: 0,
                 });
                 useListener("update-fixed-discount", this._updateFixedDiscount);
-                useListener("update-percent-discount", this._updatePercentDiscount);
                 this.remove_fixed_discount();
-                this.remove_percent_discount();
             }
             willUnmount() {
                 // Var order = this.env.pos.get_order();
@@ -36,7 +32,6 @@ odoo.define("pos_multi_discount_global.PaymentScreen", function (require) {
             }
             toggleDiscountEnabled() {
                 // Click js_fixed_discount
-                this.disable_percent_discount_mode();
                 this.currentOrder.set_fixed_discount_enabled(
                     !this.currentOrder.get_fixed_discount_enabled()
                 );
@@ -47,65 +42,27 @@ odoo.define("pos_multi_discount_global.PaymentScreen", function (require) {
                 }
                 this.render();
             }
-            togglePercentDiscountEnabled() {
-                // Click js_percent_discount
-                this.disable_fixed_discount_mode();
-                this.currentOrder.set_percent_discount_enabled(
-                    !this.currentOrder.get_percent_discount_enabled()
-                );
-                if (this.currentOrder.get_percent_discount_enabled()) {
-                    NumberBuffer.config.triggerAtInput = "update-percent-discount";
-                } else {
-                    NumberBuffer.config.triggerAtInput = "update-selected-paymentline";
-                }
-                this.render();
-            }
             _updateFixedDiscount() {
                 if (NumberBuffer.get() === null) {
                     this.remove_fixed_discount();
                 } else {
                     var disc_value = NumberBuffer.getFloat();
-                    // this.apply_discount(0, 'reset');
-                    this.apply_discount(disc_value, "fixed");
+                    this.apply_discount(0);
+                    this.apply_discount(disc_value);
                     this.state.fixed_discount = this.env.pos.format_currency(
                         disc_value
                     );
                 }
-                var order = this.env.pos.get_order();
-                if (order.percent_discount !== null && order.percent_discount > 0){
-                    this._updatePercentDiscount(null, order.percent_discount);
-                }
             }
-            _updatePercentDiscount(event, amount) {
-                if (NumberBuffer.get() === null && amount === undefined) {
-                    this.remove_percent_discount();
-                } else {
-                    if (amount !== undefined){
-                        var disc_value = amount;
-                    } else {
-                        var disc_value = NumberBuffer.getFloat();
-                    }
-                    // this.apply_discount(0, 'reset');
-                    this.reset_percent_discount();
-                    var order = this.env.pos.get_order();
-                    var total = order.get_total_without_tax();
-                    this.state.percent_discount_amount = this.env.pos.format_currency(
-                        (disc_value * total) / 100
-                    );
-                    this.apply_discount(disc_value, "percent");
-                    this.state.percent_discount = disc_value;
-                }
-            }
-            async apply_discount(amount, disc_type) {
+            async apply_discount(amount) {
                 if (this.env.pos.config.split_fixed_discount) {
                     var order = this.env.pos.get_order();
                     var lines = order.get_orderlines();
                     var total = order.get_total_without_tax();
-                    // var current_discount = order.get_total_discount();
+                    // Var current_discount = order.get_total_discount();
                     var product = this.env.pos.db.get_product_by_id(
                         this.env.pos.config.rounding_product_id[0]
                     );
-                    var abs_discount_amount = 0;
                     if (product === undefined) {
                         await this.showPopup("ErrorPopup", {
                             title: this.env._t("No rounding product found"),
@@ -115,26 +72,36 @@ odoo.define("pos_multi_discount_global.PaymentScreen", function (require) {
                         });
                         return;
                     }
-                    if (disc_type === "fixed") {
-                        this.reset_fixed_discount();
-                        total = order.get_total_without_tax();
-                        this.process_fixed_discount(amount, order);
-                        abs_discount_amount = amount;
-                    } else if (disc_type === "percent") {
-                        this.reset_percent_discount();
-                        total = order.get_total_without_tax();
-                        this.process_percent_discount(amount, order);
-                        abs_discount_amount = (amount * total) / 100;
+                    order.fixed_discount = amount;
+                    // Var discount = (100 * amount) / (total + current_discount);
+                    var discount = (100 * amount) / total;
+                    // Truncate to 3 decimals
+                    // discount = Math.trunc(discount * 10000000) / 10000000;
+                    discount = Number(Math.round(discount+'e2')+'e-2');
+                    for (const ind in lines) {
+                        if (amount > 0) {
+                            // if (lines[ind].manual_discount > 0) {
+                            //     // Discount = Math.trunc(((100 * amount) / total) * 100) / 100;
+                            //     // discount = (100 * amount) / total;
+                            //     if (discount > 0) {
+                            //         lines[ind].fixed_discount = discount;
+                            //     }
+                            // } else {
+                                lines[ind].fixed_discount = discount;
+                            // }
+                            lines[ind].set_discount(lines[ind].manual_discount);
+                        } else {
+                            lines[ind].fixed_discount = 0;
+                            lines[ind].set_discount(lines[ind].manual_discount);
+                            // Remove existing discounts
+                            if (lines[ind].get_product() === product) {
+                                order.remove_orderline(lines[ind]);
+                            }
+                        }
                     }
                     var new_total = order.get_total_without_tax();
-                    if (
-                        abs_discount_amount > 0 &&
-                        total - abs_discount_amount !== new_total
-                    ) {
-                        var diff = round_pr(
-                            new_total - (total - abs_discount_amount),
-                            0.01
-                        );
+                    if (amount > 0 && total - amount !== new_total) {
+                        var diff = round_pr(new_total - (total - amount), 0.01);
                         if (diff !== 0) {
                             if (
                                 (diff < 0 &&
@@ -147,58 +114,6 @@ odoo.define("pos_multi_discount_global.PaymentScreen", function (require) {
                     }
                     order.select_orderline(lines[0]);
                     order.deselect_orderline();
-                }
-            }
-            reset_fixed_discount() {
-                var product = this.env.pos.db.get_product_by_id(
-                    this.env.pos.config.rounding_product_id[0]
-                );
-                var order = this.env.pos.get_order();
-                var lines = order.get_orderlines();
-                for (const ind in lines) {
-                    lines[ind].fixed_discount = 0;
-                    lines[ind].set_discount(lines[ind].manual_discount);
-                    // Remove existing discounts
-                    if (lines[ind].get_product() === product) {
-                        order.remove_orderline(lines[ind]);
-                    }
-                }
-            }
-            reset_percent_discount() {
-                var product = this.env.pos.db.get_product_by_id(
-                    this.env.pos.config.rounding_product_id[0]
-                );
-                var order = this.env.pos.get_order();
-                var lines = order.get_orderlines();
-                for (const ind in lines) {
-                    lines[ind].percent_discount = 0;
-                    lines[ind].set_discount(lines[ind].manual_discount);
-                    // Remove existing discounts
-                    if (lines[ind].get_product() === product) {
-                        order.remove_orderline(lines[ind]);
-                    }
-                }
-            }
-            process_fixed_discount(amount, order) {
-                var total = order.get_total_without_tax();
-                var discount = (100 * amount) / total;
-                order.fixed_discount = amount;
-                discount = Number(Math.round(discount + "e2") + "e-2");
-                var lines = order.get_orderlines();
-                for (const ind in lines) {
-                    lines[ind].fixed_discount = discount;
-                    lines[ind].set_discount(lines[ind].manual_discount);
-                }
-            }
-            process_percent_discount(amount, order) {
-                order.percent_discount = amount;
-                var total = order.get_total_without_tax();
-                var discount = Number(Math.round(amount + "e2") + "e-2");
-                order.percent_discount_amount = (amount * total) / 100;
-                var lines = order.get_orderlines();
-                for (const ind in lines) {
-                    lines[ind].percent_discount = discount;
-                    lines[ind].set_discount(lines[ind].manual_discount);
                 }
             }
             async apply_discount_with_product(amount) {
@@ -246,26 +161,16 @@ odoo.define("pos_multi_discount_global.PaymentScreen", function (require) {
                 return super.addNewPaymentLine(...arguments);
             }
             remove_fixed_discount() {
-                // This.apply_discount(0);
+                this.apply_discount(0);
                 this.disable_fixed_discount_mode();
-            }
-            remove_percent_discount() {
-                // This.apply_discount(0);
-                this.disable_percent_discount_mode();
             }
             disable_fixed_discount_mode() {
                 this.currentOrder.set_fixed_discount_enabled(false);
                 NumberBuffer.config.triggerAtInput = "update-selected-paymentline";
-                NumberBuffer.reset();
-            }
-            disable_percent_discount_mode() {
-                this.currentOrder.set_percent_discount_enabled(false);
-                NumberBuffer.config.triggerAtInput = "update-selected-paymentline";
-                NumberBuffer.reset();
             }
         };
 
-    Registries.Component.extend(PaymentScreen, MultiDiscountPaymentScreen);
+    Registries.Component.extend(PaymentScreen, FixedDiscountPaymentScreen);
 
     return PaymentScreen;
 });
