+++ pos_multi_discount_global/static/src/js/models.js
@@ -9,15 +9,12 @@ odoo.define("pos_multi_discount_global.POSModels", function (require) {
         initialize: function (attr, options) {
             _super_orderline.initialize.call(this, attr, options);
             this.fixed_discount = 0;
-            this.percent_discount = 0;
         },
         init_from_JSON: function (json) {
             _super_orderline.init_from_JSON.apply(this, arguments);
             this.fixed_discount = json.fixed_discount;
-            this.percent_discount = json.percent_discount;
         },
         set_discount: function (discount) {
-            // Fixed_discount comes here as percent
             var parsed_discount =
                 typeof discount === "number"
                     ? discount
@@ -28,43 +25,16 @@ odoo.define("pos_multi_discount_global.POSModels", function (require) {
             this.manual_discount = disc;
             // This.discount = this.manual_discount + this.fixed_discount; // + fixed disc + % disc
             this.discount = 0;
-            if (
-                this.fixed_discount > 0 &&
-                this.percent_discount > 0 &&
-                this.manual_discount > 0
-            ) {
-                discount =
-                    100 -
-                    (1 - this.manual_discount / 100) *
-                        (1 - this.fixed_discount / 100) *
-                        (1 - this.percent_discount / 100) *
-                        100;
-                this.discount = this.round_dec(discount);
-            } else if (this.fixed_discount > 0 && this.manual_discount > 0) {
+            if (this.fixed_discount > 0 && this.manual_discount > 0) {
                 discount =
                     100 -
                     (1 - this.manual_discount / 100) *
                         (1 - this.fixed_discount / 100) *
                         100;
-                this.discount = this.round_dec(discount);
-            } else if (this.percent_discount > 0 && this.manual_discount > 0) {
-                discount =
-                    100 -
-                    (1 - this.manual_discount / 100) *
-                        (1 - this.percent_discount / 100) *
-                        100;
-                this.discount = this.round_dec(discount);
-            } else if (this.percent_discount > 0 && this.fixed_discount > 0) {
-                discount =
-                    100 -
-                    (1 - this.percent_discount / 100) *
-                        (1 - this.fixed_discount / 100) *
-                        100;
-                this.discount = this.round_dec(discount);
+                this.discount = Number(Math.round(discount + 'e2')+'e-2');
             } else if (this.fixed_discount > 0) {
-                this.discount = this.round_dec(this.fixed_discount);
-            } else if (this.percent_discount > 0) {
-                this.discount = this.round_dec(this.percent_discount);
+                discount = this.fixed_discount;
+                this.discount = Number(Math.round(discount + 'e2')+'e-2');
             } else if (this.manual_discount > 0) {
                 this.discount = this.manual_discount;
             }
@@ -74,17 +44,10 @@ odoo.define("pos_multi_discount_global.POSModels", function (require) {
         get_fixed_discount: function () {
             return this.fixed_discount;
         },
-        get_percent_discount: function () {
-            return this.percent_discount;
-        },
         export_as_JSON: function () {
             var vals = _super_orderline.export_as_JSON.apply(this, arguments);
             vals.fixed_discount = this.fixed_discount;
-            vals.percent_discount = this.percent_discount;
             return vals;
-        },
-        round_dec: function (val) {
-            return Number(Math.round(val + "e2") + "e-2");
         },
     });
 
@@ -93,39 +56,23 @@ odoo.define("pos_multi_discount_global.POSModels", function (require) {
         initialize: function (attr, options) {
             _super_order.initialize.call(this, attr, options);
             this.fixed_discount = 0;
-            this.percent_discount = 0;
-            this.percent_discount_amount = 0;
             this.fixed_discount_enabled = false;
-            this.percent_discount_enabled = false;
             this.save_to_db();
             return this;
         },
         get_fixed_discount: function () {
             return this.fixed_discount;
         },
-        get_percent_discount: function () {
-            return this.percent_discount;
-        },
-        get_percent_discount_amount: function () {
-            return this.percent_discount_amount;
-        },
         set_fixed_discount_enabled: function (fixed_discount_enabled) {
+            // This.assert_editable();
             this.fixed_discount_enabled = fixed_discount_enabled;
         },
         get_fixed_discount_enabled: function () {
             return this.fixed_discount_enabled;
         },
-        set_percent_discount_enabled: function (percent_discount_enabled) {
-            this.percent_discount_enabled = percent_discount_enabled;
-        },
-        get_percent_discount_enabled: function () {
-            return this.percent_discount_enabled;
-        },
         export_as_JSON: function () {
             var res = _super_order.export_as_JSON.apply(this, arguments);
             res.total_fixed_discount = this.fixed_discount;
-            res.total_percent_discount = this.percent_discount;
-            res.total_percent_discount_amount = this.percent_discount_amount;
             return res;
         },
     });
