+++ pos_multi_discount_global/static/src/xml/Screens/PaymentScreen/PaymentScreen.xml
@@ -13,15 +13,6 @@
                     <i class="fa fa-tag" />
                     Discount (fixed)
                 </div>
-                <div
-                    t-if="env.pos.config.split_fixed_discount"
-                    class="button js_percent_discount"
-                    t-att-class="{ highlight: currentOrder.get_percent_discount_enabled() }"
-                    t-on-click="togglePercentDiscountEnabled"
-                >
-                    <i class="fa fa-tag" />
-                    Discount (percentage)
-                </div>
             </div>
         </xpath>
         <xpath
@@ -31,8 +22,6 @@
             <PaymentScreenStatus
                 paymentLines="paymentLines"
                 fixed_discount="state.fixed_discount"
-                percent_discount="state.percent_discount"
-                percent_discount_amount="state.percent_discount_amount"
             />
         </xpath>
     </t>
