+++ pos_multi_discount_global/static/src/xml/Screens/ProductScreen/OrderWidget.xml
@@ -7,7 +7,6 @@
                 total="state.total"
                 tax="state.tax"
                 fixed_discount="state.fixed_discount"
-                percent_discount="state.percent_discount"
                 subtotal="state.subtotal"
             />
         </xpath>
